Reorganise runMainAndInitializeDB.
